Move PullList to own file
